+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Change log
 ## 1.0-SNAPSHOT
-Add user authentication with [stormpath](http://stormpath.com)
+### Added
+* `/` return the app name, description and version
+* `/version` return the version of the web service
+* User authentication with [stormpath](http://stormpath.com)
+* `/poll/create?choices=1,2,3` Create a new poll (authenticated users only)